Moved player.rs into a new "players" subpackage, to allow for future AI addition
